Migrate check plugins
